Refactor imported data
The main reason for this was the difference on
having classNames either being example Recipe_name
or Desc_name.

Recipe_name reference the specific Recipe class while
Desc_name can either be ItemDescriptor or a ResourceDescriptor.

- Remove the RecipeDocument
- Introduce ItemDescriptor

The ItemDescriptor describes all possible items
so not including resources.

But this item contain all relevent recipes and their
ingredients to easier find the information.

Also the ingredients can be either be a part or a raw resources
indicated by the "isResource" flag